% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst-tbl.R
\name{fst_tbl}
\alias{fst_tbl}
\title{Create a tbl from an fst file}
\usage{
fst_tbl(path)
}
\arguments{
\item{path}{Path to the fst file}
}
\value{
A tbl with an fst source
}
\description{
A convenience wrapper to create a table from a single-file source
}
