% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemis-package.R
\docType{package}
\name{lemis-package}
\alias{lemis}
\alias{lemis-package}
\title{lemis: The LEMIS Wildlife Trade Database}
\description{
Provides cleaned data and metadata for the US Fish and Wildlife Service's Law Enforcement Management Information System (LEMIS) data on wildlife and wildlife product imports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecohealthalliance/lemis#readme}
  \item \url{https://github.com/ecohealthalliance/lemis/tree/master/data-raw#readme}
  \item Report bugs at \url{https://github.com/ecohealthalliance/lemis/issues}
}

}
\author{
\strong{Maintainer}: Noam Ross \email{ross@ecohealthalliance.org}

Authors:
\itemize{
  \item Evan A. Eskew \email{eskew@ecohealthalliance.org}
  \item Allison M. White \email{white@ecohealthalliance.org}
  \item Carlos Zambrana-Torrelio \email{zambrana@ecohealthalliance.org}
}

Other contributors:
\itemize{
  \item EcoHealth Alliance [copyright holder]
}

}
\keyword{internal}
