% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{lemis_codes}
\alias{lemis_codes}
\title{LEMIS Code Values}
\usage{
lemis_codes()
}
\value{
A tibble with field, code, code value, and whether the field is present post-2013.
}
\description{
This function returns a data frame with descriptions of all the code values
used in \code{\link[=lemis]{lemis()}}.  This is useful for lookup (see searchable table in this help file)
as well as merging with the data for more descriptive summaries.
}
\details{
There are some changes to coding pre- and post- Febrary 2013.  Most
are just slight changes in wording and in these cases we use the post-2013
description.  Some fields codes are not used post-2013 and these have \code{FALSE}
values in the \code{post_feb_2013} column.

Original as-published PDFs of codes from pre- and post-2013 can be found in this package's
\href{https://github.com/ecohealthalliance/lemis/tree/master/inst/extdata}{\code{extdata/} directory}

\if{html}{
  \Sexpr[echo=FALSE, results=rd, stage=build]{
  in_pkgdown <- any(grepl("as_html.tag_Sexpr", sapply(sys.calls(), function(a) paste(deparse(a), collapse = "\n"))))
    if(in_pkgdown) {
      mytext <- c('In RStudio, this help file includes a searchable table of values.')
    } else {
    tmp <- tempfile(fileext=".html")
      htmlwidgets::saveWidget(DT::datatable(lemis::lemis_codes(), rownames = FALSE), tmp)
      mytext <- paste('Below is a searchable version of the LEMIS codes.',
      '\\\out{<div style="width:100\%">',
         paste(stringi::stri_subset_regex(readLines(tmp), "^</?(!DOCTYPE|meta|body|html)",negate=TRUE), collapse="\n"),
      '</div>}',
      sep="\n")
    }
    mytext
  }
}
}
