% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{lemis_metadata}
\alias{lemis_metadata}
\alias{metadata}
\title{LEMIS Field Descriptions}
\usage{
lemis_metadata()
}
\value{
A tibble with field, code, code value, and whether the field is present post-2013.
}
\description{
This function returns a data frame field descriptions for \code{\link[=lemis_data]{lemis_data()}}.
}
\details{
\if{html}{
  \Sexpr[echo=FALSE, results=rd, stage=build]{
  in_pkgdown <- any(grepl("as_html.tag_Sexpr", sapply(sys.calls(), function(a) paste(deparse(a), collapse = "\n"))))
    if(in_pkgdown) {
      mytext <- lemis:::tabular(lemis::lemis_metadata())
    } else {
    tmp <- tempfile(fileext=".html")
      htmlwidgets::saveWidget(DT::datatable(lemis::lemis_codes(), rownames = FALSE, width=700), tmp)
      mytext <- paste('Below is a searchable version of the LEMIS codes.',
      '\\\out{<div style="width:100\%">',
         paste(stringi::stri_subset_regex(readLines(tmp), "^</?(!DOCTYPE|meta|body|html)",negate=TRUE), collapse="\n"),
      '</div>}',
      sep="\n")
    }
    mytext
  }
}{
  \Sexpr[echo=FALSE, results=rd, stage=build]{
     lemis:::tabular(lemis::lemis_metadata())
  }
}
}
\seealso{
\code{\link[=lemis_codes]{lemis_codes()}} \code{\link[=lemis_data]{lemis_data()}}
}
