% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{lemis_codes}
\alias{lemis_codes}
\alias{codes}
\title{LEMIS Code Values}
\usage{
lemis_codes()
}
\value{
A tibble with field, code, code value, and whether the field is present post-2013.
}
\description{
This function returns a data frame with descriptions of all the code values
used in \code{\link[=lemis_data]{lemis_data()}}.  This is useful for lookup
as well as merging with the data for more descriptive summaries.
}
\details{
USFWS made some changes in the data dictionary in Febrary 2013.  Most
are changes slight tweaks in wording and in these cases we use the post-2013
description.  Some fields codes are not used post-2013 and these have \code{FALSE}
values in the \code{post_feb_2013} column.

Original as-published PDFs of codes from pre- and post-2013 can be found in this package's
\href{https://github.com/ecohealthalliance/lemis/tree/master/inst/extdata}{\code{extdata/} directory}

\if{html}{
  \Sexpr[echo=FALSE, results=rd, stage=build]{
  in_pkgdown <- any(grepl("as_html.tag_Sexpr", sapply(sys.calls(), function(a) paste(deparse(a), collapse = "\n"))))
    if(in_pkgdown) {
      mytext <- c('In RStudio, this help file includes a searchable table of values.')
    } else {
      mytext <- lemis::rd_datatable(lemis::lemis_codes())
    }
    mytext
  }
}

\if{text,latex}{The HTML version of this help file includes a searchable table of the LEMIS codes.}
}
\seealso{
\code{\link[=lemis_metadata]{lemis_metadata()}} \code{\link[=lemis_data]{lemis_data()}}
}
