% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{lemis_metadata}
\alias{lemis_metadata}
\alias{metadata}
\title{LEMIS Field Descriptions}
\usage{
lemis_metadata()
}
\value{
A tibble with field, code, code value, and whether the field is present post-2013.
}
\description{
This function returns a data frame field descriptions for \code{\link[=lemis_data]{lemis_data()}}.
}
\details{
\if{html}{
  \Sexpr[echo=FALSE, results=rd, stage=build]{
  in_pkgdown <- any(grepl("as_html.tag_Sexpr", sapply(sys.calls(), function(a) paste(deparse(a), collapse = "\n"))))
    if(in_pkgdown) {
      mytext <- c('In RStudio, this help file includes a searchable table of values.')
    } else {
      mytext <- lemis:::rd_datatable(lemis:::lemis_metadata())
    }
    mytext
  }
}

\if{text,latex}{The HTML version of this help file includes a searchable table of the LEMIS field metadata.}
}
\seealso{
\code{\link[=lemis_codes]{lemis_codes()}} \code{\link[=lemis_data]{lemis_data()}}
}
